+++ CHANGELOG.md
@@ -10,5 +10,6 @@
 
 ### Additions and Improvements
 - Added `--p2p-static-peers-url` option to read static peers from a URL or file
+- Added node epoch and computed slot to the sync committee duties failure message for more context about the failure condition.
 
 ### Bug Fixes